refactor(guard): split native hook Python modules
