+++ .github/workflows/rust-daemon-edge-hardening.yml
@@ -12,6 +12,8 @@ on:
       - "tests/test_daemon_bounded_http.py"
       - "tests/test_rust_daemon_edge_hardening_contracts.py"
       - "ci/native_runtime/test_native_hook_client.py"
+      - "ci/native_runtime/native_hook_client_support.py"
+      - "ci/native_runtime/test_native_hook_client_transport.py"
       - ".github/workflows/rust-daemon-edge-hardening.yml"
       - "docs/guard/rust-daemon-edge-hardening-*.md"
   push:
@@ -25,6 +27,8 @@ on:
       - "tests/test_daemon_bounded_http.py"
       - "tests/test_rust_daemon_edge_hardening_contracts.py"
       - "ci/native_runtime/test_native_hook_client.py"
+      - "ci/native_runtime/native_hook_client_support.py"
+      - "ci/native_runtime/test_native_hook_client_transport.py"
       - ".github/workflows/rust-daemon-edge-hardening.yml"
   schedule:
     - cron: "37 4 * * 2,5"
@@ -98,7 +102,7 @@ jobs:
         shell: bash
         env:
           HOL_GUARD_NATIVE_BINARY: ${{ github.workspace }}/rust/target/release/hol-guard-runtime${{ runner.os == 'Windows' && '.exe' || '' }}
-        run: uv run --no-sync pytest -q ci/native_runtime/test_native_hook_client.py --tb=short
+        run: uv run --no-sync pytest -q ci/native_runtime/test_native_hook_client.py ci/native_runtime/test_native_hook_client_transport.py --tb=short
 
   low-descriptor:
     runs-on: ubuntu-24.04
