+++ src/codex_plugin_scanner/guard/native_runtime.py
@@ -21,6 +21,8 @@ from typing import Literal, cast
 
 from .codex_hook_launch_runtime import run_isolated_hook_process
 from .native_resident_client import native_resident_client_request
+from .native_response_decoder import native_error as _native_error
+from .native_response_decoder import response_from_payload as _response_from_payload
 from .native_route_receipt import record_native_hook_result
 from .native_runtime_resilience import (
     NativeRuntimeHealthSnapshot,
@@ -53,17 +55,6 @@ _INTEGRITY_FAILURE_REASONS = frozenset(
         "native_manifest_protocol_mismatch",
         "native_manifest_rule_mismatch",
         "native_manifest_build_mismatch",
-    }
-)
-_NATIVE_ERROR_CODES = frozenset(
-    {
-        "native_overloaded",
-        "native_frame_read_failed",
-        "native_request_digest_mismatch",
-        "native_request_invalid_json",
-        "native_request_too_large",
-        "native_response_encode_failed",
-        "native_runtime_panicked",
     }
 )
 
@@ -480,83 +471,6 @@ def native_runtime_status() -> NativeRuntimeStatus:
         compatible=False,
         reason="native_unavailable",
     )
-
-
-def _response_from_payload(payload: object) -> HookReviewResponse | None:
-    if not isinstance(payload, dict):
-        return None
-    if payload.get("schema") == "guard-hook-edge-result.v2":
-        if set(payload) - {
-            "schema",
-            "authority",
-            "request_id",
-            "harness",
-            "event_name",
-            "payload_kind",
-            "result",
-        }:
-            return None
-        if (
-            payload.get("authority") != "rust"
-            or payload.get("event_name") != "PostToolUse"
-            or payload.get("payload_kind") not in {"inline", "source_file_ref", "encrypted_payload_ref"}
-            or not isinstance(payload.get("harness"), str)
-            or not payload["harness"]
-            or len(payload["harness"]) > 64
-            or not isinstance(payload.get("result"), dict)
-        ):
-            return None
-        request_id = payload.get("request_id")
-        if request_id is not None and (not isinstance(request_id, str) or len(request_id) > 256):
-            return None
-        result_payload = payload.get("result")
-        if not isinstance(result_payload, dict):
-            return None
-        payload = result_payload
-    decision = payload.get("decision")
-    model_output_action = payload.get("model_output_action")
-    notice = payload.get("notice")
-    reason_code = payload.get("reason_code")
-    if decision not in {"allow", "deny"}:
-        return None
-    if model_output_action not in {
-        "allow_original",
-        "replace_with_reviewed_excerpt",
-        "block",
-        "not_applicable",
-    }:
-        return None
-    if notice not in {"none", "excerpt", "warning"} or not isinstance(reason_code, str):
-        return None
-    reason = payload.get("reason")
-    reviewed_output_sha256 = payload.get("reviewed_output_sha256")
-    reviewed_excerpt = payload.get("reviewed_excerpt")
-    policy_action = payload.get("policy_action")
-    observed_policy_action = payload.get("observed_policy_action")
-    return HookReviewResponse(
-        decision=decision,
-        reason=reason if isinstance(reason, str) else None,
-        model_output_action=model_output_action,
-        reviewed_output_sha256=(reviewed_output_sha256 if isinstance(reviewed_output_sha256, str) else None),
-        reviewed_excerpt=(reviewed_excerpt if isinstance(reviewed_excerpt, str) else None),
-        notice=notice,
-        reason_code=reason_code,
-        policy_action=policy_action if isinstance(policy_action, str) else None,
-        observed_policy_action=(observed_policy_action if isinstance(observed_policy_action, str) else None),
-        observe_mode=payload.get("observe_mode") is True,
-    )
-
-
-def _native_error(payload: object) -> str | None:
-    if not isinstance(payload, dict) or set(payload) - {"error", "retryable"}:
-        return None
-    error = payload.get("error")
-    if not isinstance(error, str) or error not in _NATIVE_ERROR_CODES:
-        return None
-    retryable = payload.get("retryable")
-    if retryable is not None and not isinstance(retryable, bool):
-        return None
-    return error
 
 
 def _capture_native_deadline(request: HookReviewRequest) -> tuple[float, int]:
